requirements.txt
aiohttp~=3.11.18
discord.py~=2.4.0
python-dotenv~=1.0.1
aiomysql